Allow setting a custom versionist configuration
When configuring a repo with a generic type we create a temporary config
file for versionist from this template:
https://github.com/balena-io/balena-versionist/blob/master/lib/repo-type-mappings/generic/versionist.conf.js
This prevents the repo to set it's own custom `versionist.conf.js` in
case it needs to modify behavior.

This commit skips using the preset config when a versionist config is
detected in the repo root.

Change-type: minor
Signed-off-by: Robert Günzler <robertg@balena.io>