yocto-based-OS-image: Modify to match fleet versioning
When releasing into a fleet a version specified in a contract,
the initial commit is M.m.p (+rev0 is omitted) and subsequent commits are of
the form M.m.p+revN where N starts at 1, as long as the contract version
remains the same.

This change uses the same convention for OS repositories so that the
commit tag and the fleet hostapp releases match.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>